Eliminadas algunas constantes innecesarias

+++ desafio_1/include/process_data.hpp
@@ -4,5 +4,5 @@
     unsigned int* loadSeedMasking(const char* nombreArchivo, uint32_t &seed, uint32_t &n_pixels);
     bool exportImage(unsigned char* pixelData, uint16_t width, uint16_t height, QString archivoSalida);
     unsigned char* loadPixels(QString input, uint16_t &width, uint16_t &height);
-    void app_img(uint16_t n);
+    void app_img(uint8_t n);
 #endif // PROCESS_DATA_HPP
